requirements.txt
openai
streamlit
streamlit-authenticator
python-dotenv
pyyaml